[Assign plugin] Use AbstractSimpleClassPredicateMatchingService
Update assign-plugin to use the predicate matching service util
(which is also used by the all-open and no-arg plugins).

I tested manually and this resolved KT-79542.
This PR does not include tests because the issue only affected
IJ analysis, and not compilation, and I am not sure how to reproduce
the issue in a testable way.

^KT-79542